#3587 -- Ask about pruning post-install.

+++ docker.sh
@@ -396,7 +396,9 @@ update() {
         docker-compose run --rm --user="azuracast" web azuracast_update "$@"
         docker-compose up -d
 
-        docker system prune -f
+        if ask "Clean up all stopped Docker containers and images to save space?" Y; then
+            docker system prune -f
+        fi
 
         echo "Update complete!"
     fi
